Refactor server instantiation
