+++ package.json
@@ -8,9 +8,9 @@
   "license": "MIT",
   "scripts": {
     "build": "webpack --mode production",
-    "start": "npm run build && node src/server/api/index.js",
+    "start": "npm run build && node src/index.js",
     "client": "webpack-dev-server --mode development --devtool inline-source-map --hot",
-    "server": "nodemon src/server/index.js",
+    "server": "nodemon src/index.js",
     "dev": "concurrently \"npm run server\" \"npm run client\""
   },
   "dependencies": {
